Create Union of Arrays
A set of integers can be represented using a 1D array of type int with the additional condition that no
value appears in the array multiple times. Assume that two sets are represented in this manner as
arrays a[] and b[] with n_a and n_b elements, respectively. Write a function that computes the union
of these two sets and stores it in a similar manner as an array c[] with n_c elements.